refactor(core): rename _helpers to _debug_helpers
